Extract "TokenStore" component
